change ViewValue path and make TodoFormData file
データクラスを用途ごとに分けるために、app/model/viewにview関連のデータクラスをまとめた。
検索がしやすいようにViewValueのファイル名をクラス名と合わせた。
TodoFormDataはview関連のデータクラスなのでmodel/view/FormData/TodoFormDataに記述した。
FormDataのmappingをするFormオブジェクトはFormDataクラスと1対1対応するためコンパニオンオブジェクトに記述。
