+++ app/lib/model/Todo.scala
@@ -48,10 +48,3 @@ object Todo {
     )
   }
 }
-
-// Todo追加画面のフォームで使用するデータ
-case class TodoFormData(
-  categoryId: Long, //todo TodoCategoryをインポートして型を指定する必要がある？
-  title:      String,
-  body:       String
-)
