+++ app/views/Home.scala.html
@@ -6,7 +6,7 @@
 @*
 Topページ
 *@
-@(vv: model.ViewValueCommon)
+@(vv: model.view.ViewValues.ViewValueCommon)
 @common.Default(vv){
   <ul>
     <li><a>カテゴリー一覧</a></li>
