+++ app/views/todo/store.scala.html
@@ -1,7 +1,7 @@
 @*
 Todo追加画面
 *@
-@import model.ViewValueTodoStore
+@import model.view.ViewValues.ViewValueTodoStore
 @(vv: ViewValueTodoStore)(implicit messageProvider: MessagesProvider,  requestHeader: RequestHeader)
 @common.Default(vv){
   @* 追加するTodoの入力フォーム *@
